Improvised code from cluttered to clean organized functions
Improved the code structure, however the implementation is same as the previous version. Finally the iris dataset now converges to final centroids, after hours of debugging. I learnt an important thing about K-Means algorithm after this implementation, that the initial condition highly matters, if initialization is not done properly, the solution might get stuck in local minimas/maximas or might not converge at all. In the previous basic implementation I was randomly populating initial centroids, however after realizing the mistake I changed the initial centroid to randomly select any value from the dataset. This solved the problem, but it took me long time to find this logical error. However it was worth the time.