Implemented stopping criteria for the algorithm
The previous version was hardcoded to perform a fix number of iterations and return values at the end of iterations. However, now the algorithm will check if stopping condition is met, viz. if the distance table or the cluster table is not changing through few past iterations, than the algorithm will stop futher iterations and return those values, as we assume that if there are no more changes, than the solution has converged. Also added comments inside all functions for future reference.